backend/requirements.txt
Flask==2.3.3
pymongo==4.5.0
sentence-transformers==2.7.0
scikit-learn==1.3.0
numpy==1.24.3
faiss-cpu==1.8.0
redis==4.6.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0
psutil==5.9.0
beautifulsoup4==4.12.0

